+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Support for sponsored transactions
+
+### Changed
+
+- Now, the current account is selected by default for a WalletConnect session proposal,
+  instead of the one with the highest balance
+
 ## [1.17.0] - 2026-01-20
 
 ### Changed